refactor: 도메인 종속 util 메서드 분리
기존 음식점 리스트component에 있던 filter, sort메서드는 외부에서 사용되므로
캡슐화 관점에서 좋지 않았음
개선 후 domain/utils.ts에 모두 분리하였음.
